Make exepack::HEADER_LEN a u16.
It's nicer to do an infallible conversion in the u16→usize direction,
than a fallible one in the usize→u16 direction.

+++ tests/pack.rs
@@ -120,7 +120,7 @@ fn test_lengths() {
     }
 
     // The size of an EXEPACK block with no relocations.
-    let exepack_base_size = exepack::HEADER_LEN + exepack::STUB.len() + 32;
+    let exepack_base_size = usize::from(exepack::HEADER_LEN) + exepack::STUB.len() + 32;
 
     // Maximum size compressible inputs.
